LayoutTests/media/video-main-content-deny-too-small.html

<!DOCTYPE html>
<html>
<head>
    <title>video-main-content-allow</title>
    <script src="video-test.js"></script>
    <script src="media-file.js"></script>
    <script>
    function go() {
        video = document.createElement('video');
        run('internals.setMediaElementRestrictions(video, "RequireUserGestureForVideoRateChange,OverrideUserGestureRequirementForMainContent")');
        document.body.appendChild(video);

        video.src = findMediaFile('video', 'content/test');
        waitForEvent('canplaythrough', canPlayThrough);
    }

    function canPlayThrough() {
        video.play();
        waitForEventAndFail('playing');
        setTimeout(didNotBeginPlaying, 100);
    }

    function didNotBeginPlaying() {
        logResult(true, 'Did not begin playing');
        endTest();
    }
    </script>
    <style>
    video {
        width: 200px;
        height: 100px;
    }
    </style>
</head>
<body onload="go()">
</body>
</html>
