LayoutTests/media/video-main-content-deny-too-small-expected.txt

RUN(internals.setMediaElementRestrictions(video, "RequireUserGestureForVideoRateChange,OverrideUserGestureRequirementForMainContent"))
EVENT(canplaythrough)
Did not begin playing OK
END OF TEST
